move display icon folder
