+++ Backoffice/DisplayIcon/DisplayManager.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenOrchestra\BackofficeBundle\DisplayIcon;
+namespace OpenOrchestra\Backoffice\DisplayIcon;
 
 use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerAwareTrait;
