+++ Backoffice/Tests/Generator/files/Resources/views/Block/Test/show.html.twig
@@ -1,1 +1,1 @@
-This file is to modify in the OpenOrchestra\DisplayBundle bundle and the Block/Test folder
+This file is to modify in the OpenOrchestra\DisplayBundle folder and the Block/Test folder
