+++ Backoffice/Tests/Generator/files/Resources/views/Block/Test/showIcon.html.twig
@@ -1,1 +1,1 @@
-This file is to modify in the OpenOrchestra\BackofficeBundle bundle and the Block/Test folder
+This file is to modify in the OpenOrchestra\Backoffice folder and the Block/Test folder
