+++ Backoffice/Tests/Generator/files/skeleton/displayIcon/Strategy.php.twig
@@ -2,7 +2,7 @@
 
 namespace {{ namespace }}\DisplayIcon\Strategies;
 
-use OpenOrchestra\BackofficeBundle\DisplayIcon\Strategies\AbstractStrategy;
+use OpenOrchestra\Backoffice\DisplayIcon\Strategies\AbstractStrategy;
 
 /**
  * Class {{ className }}Strategy
