+++ Backoffice/Tests/Generator/files/skeleton/displayIcon/showIcon.html.twig.twig
@@ -1,1 +1,1 @@
-This file is to modify in the {{ namespace }} bundle and the Block/{{ className }} folder
+This file is to modify in the {{ namespace }} folder and the Block/{{ className }} folder
