+++ Backoffice/Tests/Manipulator/files/references/icon.yml
@@ -1,5 +1,5 @@
 services:
     open_orchestra_backoffice.icon.test:
-        class: OpenOrchestra\BackofficeBundle\DisplayIcon\Strategies\TestStrategy
+        class: OpenOrchestra\Backoffice\DisplayIcon\Strategies\TestStrategy
         tags:
             - { name: open_orchestra_backoffice.display_icon.strategy }
