+++ BackofficeBundle/Resources/skeleton/displayIcon/Strategy.php.twig
@@ -3,7 +3,7 @@
 namespace {{ namespace }}\DisplayIcon\Strategies;
 
 use OpenOrchestra\DisplayBundle\DisplayBlock\DisplayBlockInterface;
-use OpenOrchestra\BackofficeBundle\DisplayIcon\Strategies\AbstractStrategy;
+use OpenOrchestra\Backoffice\DisplayIcon\Strategies\AbstractStrategy;
 
 /**
  * Class {{ className }}Strategy
